// apps/user-service/src/message-template/placeholder.enum.ts
export enum TemplatePlaceholder {
  RH_NAME = '~#rh_name',
  RH_FIRST_NAME = '~#rh_first_name',
  RH_LAST_NAME = '~#rh_last_name',
  JOB_TITLE = '~#job_title',
  COMPANY_NAME = '~#company_name',
  SENDER_NAME = '~#sender_name',
  SENDER_FIRST_NAME = '~#sender_first_name',
  SENDER_LAST_NAME = '~#sender_last_name',
  SENDER_EMAIL = '~#sender_email',
  SENDER_PHONE = '~#sender_phone',
  SENDER_SIGNATURE = '~#sender_signature',
  JOB_LOCATION = '~#job_location',
  JOB_SALARY = '~#job_salary',
  APPLICATION_DATE = '~#application_date',
  CURRENT_DATE = '~#current_date',
}

export const PLACEHOLDER_DESCRIPTIONS = {
  [TemplatePlaceholder.RH_NAME]: 'Recruiter full name',
  [TemplatePlaceholder.RH_FIRST_NAME]: 'Recruiter first name',
  [TemplatePlaceholder.RH_LAST_NAME]: 'Recruiter last name',
  [TemplatePlaceholder.JOB_TITLE]: 'Job position title',
  [TemplatePlaceholder.COMPANY_NAME]: 'Company name',
  [TemplatePlaceholder.SENDER_NAME]: 'Your full name',
  [TemplatePlaceholder.SENDER_FIRST_NAME]: 'Your first name',
  [TemplatePlaceholder.SENDER_LAST_NAME]: 'Your last name',
  [TemplatePlaceholder.SENDER_EMAIL]: 'Your email address',
  [TemplatePlaceholder.SENDER_PHONE]: 'Your phone number',
  [TemplatePlaceholder.SENDER_SIGNATURE]: 'Your signature',
  [TemplatePlaceholder.JOB_LOCATION]: 'Job location',
  [TemplatePlaceholder.JOB_SALARY]: 'Job salary range',
  [TemplatePlaceholder.APPLICATION_DATE]: 'Application date',
  [TemplatePlaceholder.CURRENT_DATE]: 'Current date',
};